// bridge/typescript/src/_generated/dir/symbol/scope.ts
export type ScopeKind = "module" | "global" | "namespace" | "function" | "type" | "typeConditional" | "label" | "block";

export const ScopeKind = {
    /** Encode this value. */
    encode(writer: BinaryWriter, value: ScopeKind): void {
        encodeScopeKind(writer, value);
    },

    /** Decode one ScopeKind. */
    decode(reader: BinaryReader): ScopeKind {
        return decodeScopeKind(reader);
    },

    /** Return this value as JSON. */
    toJson(value: ScopeKind): Json {
        return toJsonScopeKind(value);
    },

    /** Return one ScopeKind from one JSON value. */
    fromJson(value: Json): ScopeKind {
        return fromJsonScopeKind(value);
    },
};